# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Code Graph Knowledge Service",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/sources/validate": {
      "post": {
        "tags": [
          "knowledge-sources"
        ],
        "summary": "验证知识源连接",
        "description": "校验 Git 仓库凭据及可访问分支列表。",
        "operationId": "validate_knowledge_source_connection_api_v1_admin_sources_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeSourceValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "验证结果返回",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeSourceValidationResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数不合法"
          },
          "403": {
            "description": "RBAC：权限不足"
          },
          "500": {
            "description": "内部服务器错误"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/sources": {
      "post": {
        "tags": [
          "knowledge-sources"
        ],
        "summary": "创建知识源",
        "description": "创建新的知识源记录，并立即返回规范化响应结构。",
        "operationId": "create_knowledge_source_api_v1_admin_sources_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeSourceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "知识源创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeSourceResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数不合法或知识源名称已存在"
          },
          "403": {
            "description": "RBAC：权限不足"
          },
          "500": {
            "description": "内部服务器错误"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "knowledge-sources"
        ],
        "summary": "分页获取知识源列表",
        "description": "支持分页、状态、类型与关键字搜索的知识源管理列表。",
        "operationId": "list_knowledge_sources_api_v1_admin_sources_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "页码",
              "default": 1,
              "title": "Page"
            },
            "description": "页码"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量",
              "default": 20,
              "title": "Size"
            },
            "description": "每页数量"
          },
          {
            "name": "source_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按源类型过滤",
              "title": "Source Type"
            },
            "description": "按源类型过滤"
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按激活状态过滤",
              "title": "Is Active"
            },
            "description": "按激活状态过滤"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "搜索关键词",
              "title": "Search"
            },
            "description": "搜索关键词"
          },
          {
            "name": "created_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按创建者过滤",
              "title": "Created By"
            },
            "description": "按创建者过滤"
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回知识源分页列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeSourceListResponse"
                }
              }
            }
          },
          "403": {
            "description": "RBAC：权限不足"
          },
          "500": {
            "description": "内部服务器错误"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/sources/{source_id}": {
      "get": {
        "tags": [
          "knowledge-sources"
        ],
        "summary": "获取单个知识源详情",
        "description": "根据知识源ID返回完整信息，包含最近一次同步信息。",
        "operationId": "get_knowledge_source_api_v1_admin_sources__source_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回指定知识源详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeSourceResponse"
                }
              }
            }
          },
          "403": {
            "description": "RBAC：权限不足"
          },
          "404": {
            "description": "知识源不存在"
          },
          "500": {
            "description": "内部服务器错误"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "knowledge-sources"
        ],
        "summary": "更新知识源",
        "description": "支持更新名称、描述、配置、状态等字段，返回最新信息。",
        "operationId": "update_knowledge_source_api_v1_admin_sources__source_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeSourceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功并返回最新数据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeSourceResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数不合法或知识源名称冲突"
          },
          "403": {
            "description": "RBAC：权限不足"
          },
          "404": {
            "description": "知识源不存在"
          },
          "500": {
            "description": "内部服务器错误"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "knowledge-sources"
        ],
        "summary": "软删除知识源",
        "description": "将知识源标记为非活跃状态，并记录审计日志。",
        "operationId": "delete_knowledge_source_api_v1_admin_sources__source_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "删除成功，无返回体"
          },
          "403": {
            "description": "RBAC：权限不足"
          },
          "404": {
            "description": "知识源不存在"
          },
          "500": {
            "description": "内部服务器错误"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/sources/{source_id}/stats": {
      "get": {
        "tags": [
          "knowledge-sources"
        ],
        "summary": "获取知识源统计信息",
        "description": "返回指定知识源的同步情况、任务统计等信息。",
        "operationId": "get_source_statistics_api_v1_admin_sources__source_id__stats_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回知识源统计信息",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "403": {
            "description": "RBAC：权限不足"
          },
          "404": {
            "description": "知识源不存在"
          },
          "500": {
            "description": "内部服务器错误"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/sources/jobs/{job_id}": {
      "get": {
        "tags": [
          "knowledge-sources"
        ],
        "summary": "获取解析任务详情",
        "description": "根据任务ID返回解析任务的完整状态、进度与结果信息。",
        "operationId": "get_parse_job_api_v1_admin_sources_jobs__job_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回解析任务详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParseJobResponse"
                }
              }
            }
          },
          "403": {
            "description": "RBAC：权限不足"
          },
          "404": {
            "description": "解析任务不存在"
          },
          "500": {
            "description": "内部服务器错误"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/sources/{source_id}/sync": {
      "post": {
        "tags": [
          "knowledge-sources"
        ],
        "summary": "手动触发知识源同步",
        "description": "创建新的解析任务并提交到后台同步队列。",
        "operationId": "sync_knowledge_source_api_v1_admin_sources__source_id__sync_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "object",
                    "additionalProperties": true
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Sync Config"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "同步任务提交成功",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "知识源未激活或请求参数不合法"
          },
          "403": {
            "description": "RBAC：权限不足"
          },
          "404": {
            "description": "知识源不存在"
          },
          "500": {
            "description": "内部服务器错误"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KnowledgeSourceValidationRequest": {
        "properties": {
          "repo_url": {
            "type": "string",
            "title": "Repo Url",
            "description": "Git 仓库地址"
          },
          "auth_type": {
            "type": "string",
            "enum": [
              "token",
              "none"
            ],
            "title": "Auth Type",
            "description": "认证方式",
            "default": "none"
          },
          "access_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Token",
            "description": "访问令牌（按需提供）"
          },
          "branch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch",
            "description": "待验证的分支名称"
          }
        },
        "type": "object",
        "required": [
          "repo_url"
        ],
        "title": "KnowledgeSourceValidationRequest",
        "description": "知识源连接验证请求。"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}